Utilize the logger from salesforce/core instead of pino and format the logs to logfmt
