+++ package.json
@@ -22,11 +22,12 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@salesforce/core": "^2.27.2",
+    "@salesforce/bunyan": "^2.0.0",
+    "@salesforce/core": "^3.4.2",
     "cloudevents": "^4.0.3",
     "fastify": "^3.20.2",
-    "jsforce": "^2.0.0-beta.8",
-    "pino": "^6.13.1",
+    "jsforce": "^1.10.0",
+    "logfmt": "^1.3.2",
     "whatwg-mimetype": "^2.3.0",
     "yargs": "^17.1.1"
   },
@@ -35,13 +36,12 @@
     "@types/chai": "^4.2.21",
     "@types/mocha": "^9.0.0",
     "@types/node": "^16.7.1",
-    "@types/pino": "^6.3.11",
     "@types/sinon": "^10.0.0",
     "@types/yargs": "^17.0.2",
     "@typescript-eslint/eslint-plugin": "^4.29.3",
     "@typescript-eslint/parser": "^4.29.3",
-    "concurrently": "^6.2.1",
     "chai": "^4.3.4",
+    "concurrently": "^6.2.1",
     "eslint": "^7.32.0",
     "mocha": "^9.1.0",
     "prettier": "^2.3.2",
