+++ src/sdk/data-api.ts
@@ -1,4 +1,4 @@
-import { Connection } from "jsforce/lib/connection.js";
+import Connection from "jsforce/lib/connection.js";
 import { readFileSync } from "fs";
 import { join } from "path";
 import { fileURLToPath } from "url";
